fix flipping

+++ client/src/components/Flashcard/index.jsx
@@ -8,19 +8,25 @@ function Flashcard({ id, question, answer, deleteFlashcard} ) {
 
     function handleClick(e) {
       e.preventDefault()
-      if (e.target.className === 'deleteBtn' || 'favouriteBtn') {
+      if (e.target.className === 'deleteBtn' ) {
+        return
+      } else if (e.target.className === 'favouriteBtn') {
         return
       }
-    
+ 
       setFlipped(!flipped)
     }
 
     function handleDelete(e) {
-      e.preventDefault()
-      deleteFlashcard(id)
+      const result = confirm("Are you sure you want to delete this flashcard?");
+      if (result) {
+        e.preventDefault()
+
+        deleteFlashcard(id)
+      }
+      
     }
 
-  
     function handleFavourite(e) {
       setFavourite(!favourite)
       
@@ -38,7 +44,7 @@ function Flashcard({ id, question, answer, deleteFlashcard} ) {
         </h2>
         <button className="deleteBtn" onClick={handleDelete}>X</button>
         {
-          !favourite ? <button className="favouriteBtn" onClick={handleFavourite}> &#9825;</button> : <button onClick={handleFavourite}>&hearts;</button>
+          !favourite ? <button className="favouriteBtn" onClick={handleFavourite}> &#9825;</button> : <button className="favouriteBtn" onClick={handleFavourite}>&hearts;</button>
         }
